(update) Park.jsx and Sidebar.jsx dynamically creating parks now work!
